requirements.txt
djangorestframework==2.3.13
django==1.5.1
Pygments==1.5
Markdown==2.2.0

# heroku
# run these commands to setup heroku properly
# heroku create --stack cedar --buildpack git://github.com/heroku/heroku-buildpack-python.git
# heroku config:add BUILDPACK_URL=git@github.com:heroku/heroku-buildpack-python.git#purge
# heroku config:set HEROKU=1
psycopg2==2.4.5
dj-database-url==0.2.0
